claer-cache, move extra functions utility
