+++ src/apis/cache.rs
@@ -1,40 +1,5 @@
 // Discussion: https://github.com/FifthTry/fpm/discussions/475
-// Docs: TODO
-
-/// Remove path: It can be directory or file
-async fn remove(path: &camino::Utf8PathBuf) -> std::io::Result<()> {
-    if path.is_file() {
-        tokio::fs::remove_file(path).await?;
-    } else if path.is_dir() {
-        tokio::fs::remove_dir_all(path).await?
-    } else if path.is_symlink() {
-        // TODO:
-        // It can be a directory or a file
-    }
-    Ok(())
-}
-
-/// Remove from provided `root` except given list
-async fn remove_except(root: &camino::Utf8PathBuf, except: &[&str]) -> fpm::Result<()> {
-    use itertools::Itertools;
-    let except = except
-        .iter()
-        .map(|x| root.join(x))
-        .map(|x| x.into_std_path_buf())
-        .collect_vec();
-    let mut all = tokio::fs::read_dir(root).await?;
-    while let Some(file) = all.next_entry().await? {
-        if except.contains(&file.path()) {
-            continue;
-        }
-        if file.metadata().await?.is_dir() {
-            tokio::fs::remove_dir_all(file.path()).await?;
-        } else if file.metadata().await?.is_file() {
-            tokio::fs::remove_file(file.path()).await?;
-        }
-    }
-    Ok(())
-}
+// Docs: https://fpm.dev/sitemap/clear-cache/
 
 #[derive(Default, Debug)]
 pub struct QueryParams {
@@ -43,20 +8,9 @@ pub struct QueryParams {
     all_dependencies: bool,
 }
 
-/// /api/?a=1&b=2&c=3 => vec[(a, 1), (b, 2), (c, 3)]
-fn query(uri: &str) -> fpm::Result<Vec<(String, String)>> {
-    use itertools::Itertools;
-    Ok(
-        url::Url::parse(format!("https://fifthtry.com/{}", uri).as_str())?
-            .query_pairs()
-            .into_owned()
-            .collect_vec(),
-    )
-}
-
 fn clear_cache_query(uri: &str) -> fpm::Result<QueryParams> {
     use itertools::Itertools;
-    let query = query(uri)?;
+    let query = fpm::utils::query(uri)?;
     Ok(QueryParams {
         file: query
             .iter()
@@ -117,9 +71,9 @@ pub async fn clear_(query: &QueryParams) -> fpm::Result<()> {
         let main_file_path = config.root.join(file.as_str());
         let package_file_path = config.packages_root.join(file.as_str());
         if main_file_path.exists() {
-            remove(&main_file_path).await?;
+            fpm::utils::remove(&main_file_path).await?;
         } else if package_file_path.exists() {
-            remove(&package_file_path).await?;
+            fpm::utils::remove(&package_file_path).await?;
         } else {
             println!("Not able to remove file from cache: {}", file);
         }
@@ -129,9 +83,9 @@ pub async fn clear_(query: &QueryParams) -> fpm::Result<()> {
     for package in query.package.iter() {
         if package.eq("main") {
             // TODO: List directories and files other than main
-            remove_except(&config.root, &[".packages", ".build"]).await?;
+            fpm::utils::remove_except(&config.root, &[".packages", ".build"]).await?;
         } else {
-            remove(&config.packages_root.join(package)).await?;
+            fpm::utils::remove(&config.packages_root.join(package)).await?;
         }
     }
 
@@ -145,19 +99,4 @@ pub async fn clear_(query: &QueryParams) -> fpm::Result<()> {
     }
 
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn query() {
-        assert_eq!(
-            super::query("/api/?a=1&b=2&c=3").unwrap(),
-            vec![
-                ("a".to_string(), "1".to_string()),
-                ("b".to_string(), "2".to_string()),
-                ("c".to_string(), "3".to_string())
-            ]
-        )
-    }
 }
